requirements.txt
streamlit
fastapi
uvicorn
pandas
openpyxl
customtkinter
packaging
google-genai
python-dotenv
plotly
requests
orjson
cachetools
